enhance: fold the runner group inputs into the Actions settings module
Both register global init funcs for Actions settings forms rendered across
repo, owner and admin scope, so they no longer need a file each.
