+++ web_src/js/index.ts
@@ -62,8 +62,7 @@ import {initGlobalComboMarkdownEditor, initGlobalEnterQuickSubmit} from './featu
 import {initGlobalFormDirtyLeaveConfirm} from './modules/are-you-sure.ts';
 import {callInitFunctions} from './modules/init.ts';
 import {initRepoViewFileTree} from './features/repo-view-file-tree.ts';
-import {initActionsPermissionsForm} from './features/common-actions-permissions.ts';
-import {initRunnerGroupInputs} from './features/common-runner-groups.ts';
+import {initActionsSettings} from './features/common-actions-settings.ts';
 import {initRefIssueContextPopup} from './features/ref-issue.ts';
 import {initGlobalShortcut} from './modules/shortcut.ts';
 import {initDevtest} from './modules/devtest.ts';
@@ -163,8 +162,7 @@ const initPerformanceTracer = callInitFunctions([
   initOAuth2SettingsDisableCheckbox,
 
   initRepoFileView,
-  initActionsPermissionsForm,
-  initRunnerGroupInputs,
+  initActionsSettings,
 
   initDevtest,
 ]);
